slowly moving towards proper accidental handling
key signature is implemented, but there was no system yet for dealing
with adding naturals or omitting accidentals when a previous note in
the same bar already had one. This commit adds note names together with
alteration, which is a first step towards a proper handling of
accidentals in a bar.